chtml/html/testdata/webkit/adoption01.dat

#data
<a><p></a></p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): adoption-agency-1.3
#document
| <a>
| <p>
|   <a>

#data
<a>1<p>2</a>3</p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,12): adoption-agency-1.3
#document
| <a>
|   "1"
| <p>
|   <a>
|     "2"
|   "3"

#data
<a>1<button>2</a>3</button>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,17): adoption-agency-1.3
#document
| <a>
|   "1"
| <button>
|   <a>
|     "2"
|   "3"

#data
<a>1<b>2</a>3</b>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,12): adoption-agency-1.3
#document
| <a>
|   "1"
|   <b>
|     "2"
| <b>
|   "3"

#data
<a>1<div>2<div>3</a>4</div>5</div>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,20): adoption-agency-1.3
(1,20): adoption-agency-1.3
#document
| <a>
|   "1"
| <div>
|   <a>
|     "2"
|   <div>
|     <a>
|       "3"
|     "4"
|   "5"

#data
<table><a>1<p>2</a>3</p>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-implies-table-voodoo
(1,11): unexpected-character-implies-table-voodoo
(1,14): unexpected-start-tag-implies-table-voodoo
(1,15): unexpected-character-implies-table-voodoo
(1,19): unexpected-end-tag-implies-table-voodoo
(1,19): adoption-agency-1.3
(1,20): unexpected-character-implies-table-voodoo
(1,24): unexpected-end-tag-implies-table-voodoo
(1,24): eof-in-table
#document
| <a>
|   "1"
| <p>
|   <a>
|     "2"
|   "3"
| <table>

#data
<b><b><a><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): adoption-agency-1.3
(1,16): expected-closing-tag-but-got-eof
#document
| <b>
|   <b>
|     <a>
|     <p>
|       <a>

#data
<b><a><b><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): adoption-agency-1.3
(1,16): expected-closing-tag-but-got-eof
#document
| <b>
|   <a>
|     <b>
|   <b>
|     <p>
|       <a>

#data
<a><b><b><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): adoption-agency-1.3
(1,16): expected-closing-tag-but-got-eof
#document
| <a>
|   <b>
|     <b>
| <b>
|   <b>
|     <p>
|       <a>

#data
<p>1<s id="A">2<b id="B">3</p>4</s>5</b>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,30): unexpected-end-tag
(1,35): adoption-agency-1.3
#document
| <p>
|   "1"
|   <s>
|     id="A"
|     "2"
|     <b>
|       id="B"
|       "3"
| <s>
|   id="A"
|   <b>
|     id="B"
|     "4"
| <b>
|   id="B"
|   "5"

#data
<table><a>1<td>2</td>3</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-implies-table-voodoo
(1,11): unexpected-character-implies-table-voodoo
(1,15): unexpected-cell-in-table-body
(1,30): unexpected-implied-end-tag-in-table-view
#document
| <a>
|   "1"
| <a>
|   "3"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "2"

#data
<table>A<td>B</td>C</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,8): unexpected-character-implies-table-voodoo
(1,12): unexpected-cell-in-table-body
(1,22): unexpected-character-implies-table-voodoo
#document
| "AC"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "B"

#data
<a><svg><tr><input></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,23): unexpected-end-tag
(1,23): adoption-agency-1.3
#document
| <a>
|   <svg svg>
|     <svg tr>
|       <svg input>

#data
<div><a><b><div><div><div><div><div><div><div><div><div><div></a>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): adoption-agency-1.3
(1,65): expected-closing-tag-but-got-eof
#document
| <div>
|   <a>
|     <b>
|   <b>
|     <div>
|       <a>
|       <div>
|         <a>
|         <div>
|           <a>
|           <div>
|             <a>
|             <div>
|               <a>
|               <div>
|                 <a>
|                 <div>
|                   <a>
|                   <div>
|                     <a>
|                       <div>
|                         <div>

#data
<div><a><b><u><i><code><div></a>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,32): adoption-agency-1.3
(1,32): expected-closing-tag-but-got-eof
#document
| <div>
|   <a>
|     <b>
|       <u>
|         <i>
|           <code>
|   <u>
|     <i>
|       <code>
|         <div>
|           <a>

#data
<b><b><b><b>x</b></b></b></b>y
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <b>
|   <b>
|     <b>
|       <b>
|         "x"
| "y"

#data
<p><b><b><b><b><p>x
#errors
(1,3): expected-doctype-but-got-start-tag
(1,18): unexpected-end-tag
(1,19): expected-closing-tag-but-got-eof
#document
| <p>
|   <b>
|     <b>
|       <b>
|         <b>
| <p>
|   <b>
|     <b>
|       <b>
|         "x"

#data
<b><em><foo><foob><fooc><aside></b></em>
#errors
(1,35): adoption-agency-1.3
(1,40): adoption-agency-1.3
(1,40): expected-closing-tag-but-got-eof
#document
| <b>
|   <em>
|     <foo>
|       <foob>
|         <fooc>
| <aside>
|   <b>
